Add command module
